fixing imports of pexel loader
